moved utils to its own folder for easier managing
